backend/requirements.txt
# ProprScout Property Detective Backend Dependencies

# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Database
pymongo==4.5.0
redis==4.6.0

# AI/ML Dependencies
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76

# GeoCLIP Dependencies
transformers>=4.30.0
timm>=0.9.0
einops>=0.6.0
tqdm>=4.65.0

# GeoCLIP
# Install with: pip install git+https://github.com/VicenteVivan/geo-clip.git

# Geocoding
geopy==2.3.0
requests==2.31.0

# Cloudinary
cloudinary==1.41.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
joblib>=1.3.0
